Improve detection of reference sequence file.
Improve how the reference file is detected, avoiding the cases where
a file with a different extension than the prefix is selected, which
may lead to using the wrong file.

E.g. if the prefix is .../prefix.fst but the file .../prefix.fa
also exists, then this file would be selected even though the
correct file is .../prefix.fst.