Allows BWA to be correctly executed on OSX.
This changes the command-line call to adhere to the stricter
requirements imposed by OSX, namely that options directly follow the
command name. To accomplish this, another class of parameters is added
(so-called positionals), which are added as the last part of the
command. Typically this will be the input/output files, as exemplified
by the following (where "input.bam" is the positional):

$ samtools view -h input.bam

This allows additional options to be set, without breaking the command.