Limits contigs enumerated in bam_coverage / bam_depths.
The maximum number of contigs enumerated in .coverage / .depths
files are now restricted to 100, in order to prevent the tables
from exploding in size due to poorly assembled genomes. This
limitation is not enforced if a set of intervals are specified.

In addition, this fixes the contig size being reported incorrectly
for contigs with no hits, and the absence of the last column for
contigs with no hits.