Improve pre-commit file-state.
Checks for missing files are now done at the Node level, allowing
checks to take into consideration all files that are expected to
have been created by a set of commands. In addition, this makes it
possible detect the creation of unexpected files before other files
have been moved/removed, making it easier to disect such failures
as the entire pre-commit state is preserved when the error is
detected.