Use cPickle to test nodes during start-up.
Use cPickle instead of the python pickle module to perform the
pickling test of nodes (to ensure that these can be used with
multiprocessing). This shaves off a chunk of the start-up time.

If the test fails, the python pickle module is used to generate
the error message, since it provides a more informative message
than cPickle.