Improve handling of SIGINT.
Rather than catching KeyboardInterrupt, we now set a flag to let the pipeline
know that it should terminate; this is done to prevent corruption of data-
structures. Subsequent SIGINTs are handled by the default handler, leading to
the termination of the pipeline.

Attempts to call 'terminate' on the pool of processes have been abandonded, as
this did not seem to work reliably, instead leading to the process locking up.
The default behavior appears to be adequate.