Fix merging by label in .summary files.
Previously, only the size of the one prefix would be used, if multiple
prefixes had the same label. Furthermore, all prefixes with that label
would be identical in the table of genomes. This corrects this behavior
so that the total size is used, and that the individual stats are shown
in the genomes table.