Fix rare case of spurious insertions.
This fixes a rare case where spurious insertions could be added to
sequences, if two conditions were met: Firstly, the insertion had
to start before the sequence, and secondly the resulting negative
position (indel.pos - start) had to extend past the 5' padding.

This was exposed due to the removal of 5' padding.