Minor rework of new depth histogram code.
Allows the histogram for the entire genome to be calculated from
the partial BAMs (e.g. library BAMs), while requiring that the
final BAM(s) have been built if regions are enabled. Thus, this
allows the long-running task of calculating the whole genome
histogram to be done in parallel with the creation of the final
BAM, and only the (typically) short running region specific
histograms have to be done AFTER the final BAM has been built.