Reworking sequence nodes, and related phylo parts.
Reworking / cleanup of 'sequences' nodes; add extra checks to
ensure that sequences are correctly collected; index genotyped
sequences to simplify merging.

Reference sequences included in the phylo pipeline are now expected
to be located in the 'refseqs' folder; thus if sample X is a
"Reference Sequence", the default location is expected to be
./data/refseqs/{Prefix}.{X}.fasta. This is to allow mutliple
reference sequences and/or different prefixes.