Avoid using STDOUT from java tools (Picard, etc.)
This is nessesary as the JRE may write debug information to STDOUT
when native libraries are loaded, leading to mixed / invalid output.

The message is as follows:
"You have loaded library XYZ.so which might have disabled stack guard. The VM
will try to fix the stack guard now. It's highly recommended that you fix the
library with 'execstack -c ', or link it with '-z noexecstack'."