Allow running with "too few" threads, but warn.
Allows the pipeline to run even if the maximum number of threads
specified is less than the max required by any given node. If nodes with
a higher number of threads exist, the pipeline issues a warning, and
only runs these when no other nodes are running.

This is required since the minimum is greater than 1, due to the
concurrent execution of BWA for PE reads.