Check that human genomes are GATK compatible
Add checks for chromosome ordering in human genomes (hg18, hg19, and hg38),
requiring that these be sorted in numerical order (chr1, chr2, chr3, ...)
rather than lexcially (chr1, chr10, chr11, ...). This will make the first run
of the pipeline with a new index somewhat slower (as the FAI index may have to
be created), but will potentially save a lot of work downstream, by not
requiring that the BAM be fixed later.