Fix new FAI files being outdated in first run.
Due to changes in the pipeline, which now opens / indices FASTA files
during startup, the resulting FAI would be considered out-of-date due
to the dependency on the FASTA validation step. The index step no
longer depends on the validation step, under the assumption that most
FASTA files are valid, and hence little is lost by greedy re-indexing.