Rework 'sample_pileup' command to reduce mem use.
Note that this also changes the meaning of the minimum distance to
indels to be more sensible; distance is now measured from the location
of the indel. For example, for a deletion D, distance are as follows:

Pileup:    . . . . . - - . . . . .
Distance:  5 4 3 2 1 0 0 1 2 3 4 5

Thus, the site containing deleted has a distance of 0, while the adjacent
bases have a distance of 1, contrary to previously where adjacent bases
had a distance of 0. Thus, to exclude only sites containing deleted
bases, a min. distance of 1 may be used, while a min. distance of 0
filters no bases.

For insertions, distances are calculated similarly, except that inserted
bases can never be included, regardless of the min. distance to indels:

Pileup:    . . . . . + + . . . . .
Distance:  5 4 3 2 1     1 2 3 4 5

This is not directly mappable onto the previous behavior.