support intermediate files
This is a major rewrite of the task graph to support intermediate
files generated parts of pipelines. This may for example be .sai files
from running `bwa aln` or pre-merged BAM files.

When appropriately marked, these files are automatically removed once
tasks depending on them have finished running. This significantly
decreases the amount of disk-space needed to run a pipeline.